requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
instagrapi==2.1.3
pillow==11.1.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
